% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmas.extract.binned.R
\name{dmas.extract.binned}
\alias{dmas.extract.binned}
\title{Extract a binned model}
\usage{
dmas.extract.binned(apikey, endpoints, coeffs, infoid, id = NA)
}
\arguments{
\item{apikey}{API Key provided by DMAS}

\item{endpoints}{list of endpoints, e.g. c('-inf', .2, .4, 'inf')}

\item{coeffs}{list of coefficients, with 'drop' for the dropped
bin, e.g. c('c1', 'drop', 'c3')}

\item{infoid}{An InfoID (gcp or from the DMAS spreadsheet)}

\item{id}{ID returned by dmas.put.model (or NA for last)}
}
\description{
Identify a binned model for use in DMAS after uploading it with
dmas.put.model.
}

