% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmas.extract.polynomial.R
\name{dmas.extract.polynomial}
\alias{dmas.extract.polynomial}
\title{Extract a polynomial model}
\usage{
dmas.extract.polynomial(apikey, coeffs, lowbound, highbound, infoid, id = NA)
}
\arguments{
\item{apikey}{API Key provided by DMAS}

\item{coeffs}{list of coefficients, e.g. c('lin', 'quad', 'cub')}

\item{lowbound}{lower bound}

\item{highbound}{upper bound}

\item{infoid}{An InfoID (gcp or from the DMAS spreadsheet)}

\item{id}{ID returned by dmas.put.model (or NA for last)}
}
\description{
Identify a polynomial model for use in DMAS after uploadingit with
dmas.put.model.
}

