% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmas.extract.single.R
\name{dmas.extract.single}
\alias{dmas.extract.single}
\title{Extracting a single coefficient model}
\usage{
dmas.extract.single(apikey, coeff, infoid, id = NA)
}
\arguments{
\item{apikey}{API Key provided by DMAS}

\item{infoid}{An InfoID (gcp or from the DMAS spreadsheet)}

\item{id}{ID returned by dmas.put.model (or NA for last)}

\item{coeffs}{A coefficient name}
}
\description{
Identify a model for use in DMAS after uploading it with dmas.put.model.
}

