% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmas.get.api.R
\name{dmas.get.api}
\alias{dmas.get.api}
\title{Low-level function for accessing the API}
\usage{
dmas.get.api(methodargs, as.model, quietly = F)
}
\arguments{
\item{methodargs}{The URL string, after the server; eg. test?arg=3}

\item{as.model}{Should we interpret the result as a model ID?}

\item{quietly}{Should an OK result be silent?}
}
\description{
Clearing-house function for calling the DMAS API
}

