% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmas.put.model.R
\name{dmas.put.model}
\alias{dmas.put.model}
\title{Sending a model to the DMAS server}
\usage{
dmas.put.model(model, apikey, infoid, varnum = NA)
}
\arguments{
\item{model}{And LM or FELM model}

\item{apikey}{API Key provided by DMAS}

\item{infoid}{An InfoID (gcp or from the DMAS spreadsheet)}

\item{varnum}{An optional number of coefficients to upload}
}
\description{
Send all model data to DMAS
}

